Move migration/conn string out of model
I didn't realize this but there was a weird dependence on
models for defining the migration path and the connection
to the sqlite database.

This gets rid of that by moving that stuff to the database
object. This way migrations and models can be run in any order
or in different files or whatever.